Separated non clear Renderpass and densoiser creation to extra util files. Main file is reduced to under 600 lines of code.
