+++ source/util/VsgUtils.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <vsg/all.h>
+
+namespace VsgUtils{
+     vsg::ref_ptr<vsg::RenderPass> createNonClearRenderPass(VkFormat colorFormat, VkFormat depthFormat, vsg::Device* device);
+}
